requirements.txt
torch
torchvision
torchaudio
scipy
librosa
numba
tqdm
hydra-core
omegaconf
pyloudnorm
tensorboard
wandb